readline: Allow linefeeds to be used to end lines
Currently readline ignores linefeeds which works well for \r\n (and \r)
data but has problems with linefeed seperations (which, amoung other
things, is used by Gadgetbridge's BangleJS driver.

Add support for executing commands on a linefeed whilst ensuring that
we continue to treat \r\n as a single line.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>