Set ADC ACQTIME to 40us in nrf
In the pinetime, the battery is connected with a 1MOhm resistor, so 3us
isn't enough to measure it.
The battery is the only place where the ADC is used in the pinetime, so
it's ok to hardcode the new value, but it'd be better to make it
configurable if contributing upstream.

Fixes https://github.com/daniel-thompson/wasp-os/issues/254

For more details see the same issue in infinitime:
https://github.com/InfiniTimeOrg/InfiniTime/pull/483

Signed-off-by: Francesco Gazzetta <fgaz@fgaz.me>
[daniel@redfelineninja.org.uk: rebased on v1.15]
Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>